:pencil: DataReader and Common
