+++ Common/inc/logger.h
@@ -1,3 +1,12 @@
+/*
+*  FILE          : logger.h
+*  PROJECT       : Assignment #3
+*  PROGRAMMER    : Gabriel Gurgel & Michael Gordon
+*  FIRST VERSION : 2020-03-21
+*  DESCRIPTION   : Shared constants, includes, and structures for the Hoochamacallit
+*                  logger functionality
+*/
+
 #pragma once
 #include <stdio.h>
 #include <time.h>
@@ -5,6 +14,7 @@
 
 #define LOG_FOLDER_PATH "/tmp/"
 #define LOG_LEN         256
+#define TIME_LEN        50
 
 // Function Prototypes
 void getTime(char* output);
