+++ dataReader/makefile
@@ -1,12 +1,12 @@
 #
 # this makefile will compile and and all source
-# found in the ~/develop/dataReader directory.
+# found in the dataReader directory.
 #
 #==============================================
 
 # FINAL BINARY Target
-./bin/DataReader : ./obj/dataReader.o ../Common/obj/logger.o  ./obj/DClist.o  ./obj/DCOutput.o
-	cc ./obj/dataReader.o ../Common/obj/logger.o  ./obj/DClist.o ./obj/DCOutput.o -o  ./bin/DataReader
+./bin/DataReader : ./obj/dataReader.o ../Common/obj/logger.o  ./obj/DClist.o  ./obj/DROutput.o
+	cc ./obj/dataReader.o ../Common/obj/logger.o  ./obj/DClist.o ./obj/DROutput.o -o  ./bin/DataReader
 #
 # =======================================================
 #                     Dependencies
@@ -15,8 +15,8 @@
 ./obj/dataReader.o : ./src/dataReader.c  ./inc/dataReader.h
 	cc -c ./src/dataReader.c -o  ./obj/dataReader.o
 
-./obj/DCOutput.o : ./src/DCOutput.c  ./inc/dataReader.h
-	cc -c ./src/DCOutput.c -o  ./obj/DCOutput.o
+./obj/DROutput.o : ./src/DROutput.c  ./inc/dataReader.h
+	cc -c ./src/DROutput.c -o  ./obj/DROutput.o
 
 ./obj/DClist.o : ./src/DClist.c ./inc/dataReader.h
 	cc -c ./src/DClist.c -o ./obj/DClist.o
