+++ dataReader/src/dataReader.c
@@ -1,10 +1,17 @@
+/*
+*  FILE          : dataReader.c
+*  PROJECT       : Assignment #3
+*  PROGRAMMER    : Gabriel Gurgel & Michael Gordon
+*  FIRST VERSION : 2020-03-21
+*  DESCRIPTION   : DataReader Application, The Server of the Hoochamacallit system
+*                  Creates the semaphores to the client apps, creates the message queue
+*                  and the shared memory section
+*/
+
 #include "../inc/dataReader.h"
 
 // DEBUG: REMOVE ALL PRINTF
 // DEBUG: REMOVE THE DEBUGGER FLAG
-DCInfo createClient(pid_t id);
-int setUpLogSemaphore(void);
-void closeLogSemaphore(int semId);
 
 struct sembuf acquire_operation = { 0, -1, SEM_UNDO };
 struct sembuf release_operation = { 0, 1, SEM_UNDO };
@@ -60,13 +67,6 @@ int main(int argc, char const *argv[])
   }
   printf("DEBUG SHM ID: %d\n", shmID);
 
-  /*
-  //DEBUG: Clean up
-  msgctl (msgID, IPC_RMID, (struct msqid_ds*)NULL);
-  shmdt(shList);
-  shmctl (shmID, IPC_RMID, 0);
-  */
-
   //--> Listening loop
   shList = (MasterList*)shmat (shmID, NULL, 0);       // Grabs the shared memory and
   shList->msgQueueID = msgID;                         // Assign the message queue ID
@@ -77,8 +77,8 @@ int main(int argc, char const *argv[])
   msgData msg;
   int msgSize = sizeof(msgData) - sizeof(long);
 
-  time_t startTime = time(NULL);                      // Listen for messages loop
-  while((int)difftime(time(NULL), startTime) < EXIT_DELAY)
+  time_t startTime = time(NULL); // Listen for messages loop
+  while((int)difftime(time(NULL), startTime) < CLOSE_DELAY)
   {
     // Process messages if received AND if it is able to add client or it already exists in shList
     if(msgrcv(msgID, &msg, msgSize, 0, IPC_NOWAIT) != -1 &&
@@ -89,12 +89,12 @@ int main(int argc, char const *argv[])
       checkInactivity(shList);
       int currentClient = findClient(shList, msg.clientId);
 
-      if(msg.msgStatus == EXIT_CODE)
+      if(msg.msgStatus == EXIT_CODE)        // Remove client
       {
         deleteNode(shList, currentClient);
         createLogMessage(shList->dc[currentClient], GO_OFFLINE, currentClient, 0);
       }
-      else
+      else                                 // Log message
       {
         createLogMessage(shList->dc[shList->numberOfDCs], MESSAGE, shList->numberOfDCs, msg.msgStatus);
       }
@@ -109,7 +109,7 @@ int main(int argc, char const *argv[])
   debugLog("All DCs have gone offline or terminated – DR TERMINATING\n");
   msgctl (msgID, IPC_RMID, (struct msqid_ds*)NULL);
   shmdt(shList);
-  shmctl (shmID, IPC_RMID, 0);
+  shmctl(shmID, IPC_RMID, 0);
 
   return 0;
 }
@@ -140,11 +140,4 @@ int setUpLogSemaphore(void)
 void closeLogSemaphore(int semId)
 {
   semctl (semId, 0, IPC_RMID, 0);
-
-}
-
-DCInfo createClient(pid_t id)
-{
-  DCInfo client = {.dcProcessID = id, .lastTimeHeardFrom = time(NULL)};
-  return client;
 }
